ci: split tests for better parallelism
