+++ CODEOWNERS
@@ -230,6 +230,7 @@ libs/ledger-live-common/src/wallet-api/                  @ledgerhq/coin-integrat
 **/liveapp-sdk.spec.ts                                   @ledgerhq/coin-integration
 **/confirmTransaction.spec.ts                            @ledgerhq/coin-integration
 **/wallet-api.spec.ts                                    @ledgerhq/coin-integration
+apps/ledger-live-mobile/e2e/specs/wallet-api/            @ledgerhq/coin-integration
 **/wallet-api.spec.ts-snapshots/                         @ledgerhq/coin-integration
 **/dapp.spec.ts                                          @ledgerhq/coin-integration
 libs/test-utils/                                         @ledgerhq/coin-integration
