+++ apps/ledger-live-mobile/e2e/specs/deeplinks.spec.ts
@@ -1,4 +1,4 @@
-import { knownDevices } from "../models/devices";
+import { initDeeplinksMockApp } from "./deeplinksFlow.shared";
 
 $TmsLink("B2CQA-1837");
 describe("DeepLinks Tests", () => {
@@ -7,16 +7,7 @@ describe("DeepLinks Tests", () => {
   const arbitrumLong = "arbitrum";
 
   beforeAll(async () => {
-    await app.init({
-      userdata: "1AccountBTC1AccountETHReadOnlyFalse",
-      knownDevices: [knownDevices.nanoX],
-      featureFlags: {
-        noah: {
-          enabled: false,
-        },
-      },
-    });
-    await app.portfolio.waitForPortfolioPageToLoad();
+    await initDeeplinksMockApp();
   });
 
   it("should open My Ledger page", async () => {
@@ -42,18 +33,6 @@ describe("DeepLinks Tests", () => {
   it("should open Custom Lock Screen page", async () => {
     await app.customLockscreen.openViaDeeplink();
     await app.customLockscreen.expectCustomLockscreenPage();
-  });
-
-  it("should open the Discover page", async () => {
-    await app.discover.openViaDeeplink();
-    await app.discover.expectDiscoverPage();
-  });
-
-  it(`should open discovery to random live App`, async () => {
-    // Opening only one random liveApp to avoid flakiness
-    const randomLiveApp = app.discover.getRandomLiveApp();
-    await app.discover.openViaDeeplink(randomLiveApp);
-    await app.discover.expectApp(randomLiveApp);
   });
 
   it("should open Market Detail page for Bitcoin", async () => {
