--- apps/ledger-live-mobile/e2e/specs/receive/receiveFlow.spec.ts
@@ -1,100 +1,0 @@
-import DeviceAction from "../../models/DeviceAction";
-import { knownDevices } from "../../models/devices";
-
-describe("Receive Flow", () => {
-  let deviceAction: DeviceAction;
-  const btcReceiveAddress = "173ej2furpaB8mTtN5m9829MPGMD7kCgSPx";
-  let first = true;
-  const knownDevice = knownDevices.nanoX;
-
-  beforeAll(async () => {
-    await app.init({
-      userdata: "EthAccountXrpAccountReadOnlyFalse",
-      knownDevices: [knownDevice],
-      featureFlags: {
-        noah: {
-          enabled: false,
-        },
-      },
-    });
-    deviceAction = new DeviceAction(knownDevice);
-
-    await app.portfolio.waitForPortfolioPageToLoad();
-  });
-
-  async function openReceive() {
-    await app.portfolio.openViaDeeplink();
-    await app.portfolio.waitForPortfolioPageToLoad();
-    await app.transferMenu.open();
-    await app.transferMenu.navigateToReceive();
-  }
-
-  $TmsLink("B2CQA-1864");
-  it("Should verify the address after importing an account working on a single network", async () => {
-    await openReceive();
-    await app.modularDrawer.performSearchByTicker("BTC");
-    await app.modularDrawer.selectCurrencyByTicker(Currency.BTC.ticker);
-    await app.modularDrawer.tapAddNewOrExistingAccountButtonMAD();
-
-    first && (await deviceAction.selectMockDevice(), (first = false));
-    await deviceAction.openApp();
-    await app.addAccount.addAccountAtIndex(Currency.BTC.name, Currency.BTC.id, 0);
-    await app.receive.selectVerifyAddress();
-    await deviceAction.openApp();
-    await app.receive.expectAddressIsVerified(btcReceiveAddress);
-    await deviceAction.complete();
-    await app.receive.expectAddressIsDisplayed(btcReceiveAddress);
-  });
-
-  $TmsLink("B2CQA-1858");
-  $TmsLink("B2CQA-1860");
-  it("Should display the number of account existing per networks", async () => {
-    await openReceive();
-    await app.modularDrawer.performSearchByTicker("ETH");
-    await app.modularDrawer.selectCurrencyByTicker(Currency.ETH.ticker);
-    await app.modularDrawer.expectNumberOfAccountInListIsDisplayed("Ethereum", 4);
-    await app.modularDrawer.expectNumberOfAccountInListIsDisplayed("OP Mainnet", 1);
-    await app.modularDrawer.tapDrawerCloseButton();
-  });
-
-  $TmsLink("B2CQA-1856");
-  $TmsLink("B2CQA-1862");
-  it("Should create an account on a network", async () => {
-    const currency = Currency.OP;
-    await openReceive();
-    await app.modularDrawer.performSearchByTicker("ETH");
-    await app.modularDrawer.selectCurrencyByTicker(Currency.ETH.ticker);
-    await app.modularDrawer.selectNetworkIfAsked(currency.name);
-
-    await app.modularDrawer.tapAddNewOrExistingAccountButtonMAD();
-    first && (await deviceAction.selectMockDevice(), (first = false));
-    await deviceAction.openApp();
-    await app.addAccount.addAccountAtIndex(Currency.OP.name, Currency.OP.id, 2);
-    await app.receive.doNotVerifyAddress();
-    await app.receive.expectReceivePageIsDisplayed("ETH", "OP Mainnet 3");
-  });
-
-  $TmsLink("B2CQA-650");
-  it("Should access to receive after importing a cryptocurrency on a selected network", async () => {
-    await openReceive();
-    await app.modularDrawer.performSearchByTicker("ETH");
-    await app.modularDrawer.selectCurrencyByTicker(Currency.ETH.ticker);
-    await app.modularDrawer.selectNetworkIfAsked("Arbitrum");
-    await app.modularDrawer.tapAddNewOrExistingAccountButtonMAD();
-    first && (await deviceAction.selectMockDevice(), (first = false));
-    await deviceAction.openApp();
-    await app.addAccount.addAccountAtIndex("Arbitrum", "arbitrum", 0);
-    await app.receive.doNotVerifyAddress();
-    await app.receive.expectReceivePageIsDisplayed("ETH", "Arbitrum 1");
-  });
-
-  $TmsLink("B2CQA-1859");
-  it("Should access to receive after selecting an existing account", async () => {
-    await openReceive();
-    await app.modularDrawer.performSearchByTicker("XRP");
-    await app.modularDrawer.selectCurrencyByTicker(Currency.XRP.ticker);
-    await app.modularDrawer.selectAccount(1);
-    await app.receive.doNotVerifyAddress();
-    await app.receive.expectReceivePageIsDisplayed("XRP", "XRP 2");
-  });
-});
