+++ apps/ledger-live-mobile/e2e/specs/send/currencies.cosmos.spec.ts
@@ -4,23 +4,10 @@ import { CryptoCurrencyId } from "@ledgerhq/types-cryptoassets";
 import { formattedAmount, getAccountName, getAccountUnit } from "../../models/currencies";
 
 $TmsLink("B2CQA-1823");
-describe("Send flow", () => {
+describe("Send flow (cosmos)", () => {
   let deviceAction: DeviceAction;
   let first = true;
-  const testedCurrencies: CryptoCurrencyId[] = [
-    "bitcoin",
-    "ethereum",
-    "bsc",
-    //"ripple", // TOFIX Error during flow
-    //"solana", // TOFIX Error during flow
-    //"cardano", // TOFIX Error during flow
-    "dogecoin",
-    //"tron", // TOFIX Error during flow
-    //"avalanche_c_chain", // TOFIX Error during flow
-    "polygon",
-    "polkadot",
-    "cosmos",
-  ];
+  const testedCurrencies: CryptoCurrencyId[] = ["cosmos"];
   const knownDevice = knownDevices.nanoX;
 
   beforeAll(async () => {
@@ -57,7 +44,10 @@ describe("Send flow", () => {
       await app.send.expectSummaryAmount(amountWithCode);
       await app.send.summaryContinue();
 
-      first && (await deviceAction.selectMockDevice(), (first = false));
+      if (first) {
+        await deviceAction.selectMockDevice();
+        first = false;
+      }
       await deviceAction.openApp();
 
       await app.common.successClose();
