+++ apps/ledger-live-mobile/e2e/specs/wallet-api/transactionSign.todo.spec.ts
@@ -1,75 +1,9 @@
-import DeviceAction from "../models/DeviceAction";
-import { knownDevices } from "../models/devices";
+import { afterAllWalletApi, afterEachWalletApi, beforeAllWalletApi } from "./walletApiLifecycle";
 
-describe("Wallet API methods", () => {
-  const knownDevice = knownDevices.nanoX;
-  let deviceAction: DeviceAction;
-
-  beforeAll(async () => {
-    await app.init({
-      userdata: "1AccountBTC1AccountETHReadOnlyFalse",
-      knownDevices: [knownDevice],
-    });
-    await app.dummyWalletApp.startApp();
-
-    await app.portfolio.waitForPortfolioPageToLoad();
-    await app.dummyWalletApp.openApp();
-    await app.dummyWalletApp.expectApp();
-    deviceAction = new DeviceAction(knownDevice);
-  });
-
-  afterAll(async () => {
-    await app.dummyWalletApp.stopApp();
-  });
-
-  afterEach(async () => {
-    await app.dummyWalletApp.clearStates();
-  });
-
-  it("account.request", async () => {
-    await app.dummyWalletApp.sendRequest();
-    await app.cryptoDrawer.selectCurrencyFromDrawer("Bitcoin");
-    await app.cryptoDrawer.selectAccountFromDrawer("Bitcoin 1 (legacy)");
-
-    const res = await app.dummyWalletApp.getResOutput();
-    expect(res).toMatchObject({
-      id: "2d23ca2a-069e-579f-b13d-05bc706c7583",
-      address: "1xeyL26EKAAR3pStd7wEveajk4MQcrYezeJ",
-      balance: "35688397",
-      blockHeight: 194870,
-      currency: "bitcoin",
-      name: "Bitcoin 1 (legacy)",
-      spendableBalance: "35688397",
-    });
-  });
-
-  it("account.receive", async () => {
-    await app.dummyWalletApp.sendAccountReceive();
-    await app.walletAPIReceive.continueWithoutDevice();
-    await app.walletAPIReceive.cancelNoDevice();
-    await app.walletAPIReceive.continueWithoutDevice();
-    await app.walletAPIReceive.confirmNoDevice();
-
-    const res = await app.dummyWalletApp.getResOutput();
-    expect(res).toBe("1xeyL26EKAAR3pStd7wEveajk4MQcrYezeJ");
-  });
-
-  it("message.sign", async () => {
-    const account = "Ethereum 1";
-    const message = "Hello World! This is a test message for signing.";
-
-    await app.dummyWalletApp.setAccountId("e86e3bc1-49e1-53fd-a329-96ba6f1b06d3");
-    await app.dummyWalletApp.setMessage(message);
-    await app.dummyWalletApp.messageSign();
-
-    await app.walletAPISignMessage.expectSummary(account, message);
-    await app.walletAPISignMessage.summaryContinue();
-    await deviceAction.selectMockDevice();
-    await deviceAction.silentSign();
-
-    const res = await app.dummyWalletApp.getResOutput();
-    expect(res).toBe("mockedSignature");
-  });
+describe("Wallet API methods — transaction.sign (todo)", () => {
+  beforeAll(beforeAllWalletApi);
+  afterAll(afterAllWalletApi);
+  afterEach(afterEachWalletApi);
 
   xit("transaction.sign", async () => {
     const recipient = "0x046615F0862392BC5E6FB43C92AAD73DE158D235";
