+++ apps/ledger-live-mobile/package.json
@@ -27,7 +27,7 @@
     "mock:test": "export MOCK=1 && pnpm detox test",
     "e2e:test": "pnpm --filter ledger-live-mobile-e2e-tests test:detox",
     "prebeta": "bundle install",
-    "debug:detox": "pnpm detox test -c ios.manual currencies.spec.ts",
+    "debug:detox": "pnpm detox test -c ios.manual e2e/specs/send/currencies.bitcoinDogecoin.spec.ts",
     "ios:staging": "ENVFILE=.env.ios.staging react-native run-ios --mode Staging",
     "preios:local:ipa": "bundle install",
     "ios:local:ipa": "bundle exec fastlane ios local_ipa --env ios.staging",
